test_groq.py
import os
from dotenv import load_dotenv
from groq import Groq

load_dotenv()

api_key = os.getenv('GROQ_API_KEY')
print(f"API Key found: {api_key[:10]}..." if api_key else "No API key found")

if api_key:
    try:
        client = Groq(api_key=api_key)

        response = client.chat.completions.create(
            model="llama-3.3-70b-versatile",
            messages=[
                {"role": "user", "content": "Say hello in one sentence"}
            ],
            max_tokens=50
        )

        print("✓ API Key works!")
        print(f"Response: {response.choices[0].message.content}")

    except Exception as e:
        print(f"✗ API Key test failed: {e}")
else:
    print("✗ No API key found in .env file")