requirements.txt
# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0
gunicorn==21.2.0

# API Clients
google-api-python-client==2.108.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2

# RAG Components (No compilation needed!)
sentence-transformers==2.2.2
numpy==1.24.3
scikit-learn==1.3.0
groq==0.4.2